pyproject.toml
[tool.poetry]
name = "aw-leaderboard"
version = "0.1.0"
description = ""
authors = ["Erik Bjäreholt <erik@bjareho.lt>"]
readme = "README.md"
packages = [
    { include = "aw_leaderboard", from = "src" },
]

[tool.poetry.scripts]
aw-leaderboard = "aw_leaderboard.main:main"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
pydantic = "^1.10.7"
sqlalchemy = "^2.0.13"
pyjwt = {extras = ["crypto"], version = "^2.7.0"}
passlib = "^1.7.4"
asyncpg = "^0.27.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"
aiofiles = "^23.1.0"
aiosqlite = "^0.19.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
pyupgrade = "^3.4.0"
pytest = "^7.3.1"
ruff = "^0.0.267"
types-passlib = "^1.7.7.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
